controller stuff, i forgot i was suppose to work on sound
