+++ tests/pixel_format_tests.rs
@@ -1,4 +1,3 @@
-
 use beryllium::*;
 
 #[rustfmt::skip]
